requirements.txt
asn1crypto==0.24.0
cached-property==1.5.1
cachetools==2.1.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.3.1
Flask==1.0.2
google-api-core==1.5.0
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-storage==1.13.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
mmh3==2.5.1
optimizely-sdk==2.1.1
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pyOpenSSL==18.0.0
pytz==2018.5
requests==2.19.1
rsa==4.0
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
